+++ src/playmp_controller.cpp
@@ -227,7 +227,7 @@ void playmp_controller::linger()
 	if(get_end_level_data().transient.reveal_map) {
 		// Change the view of all players and observers
 		// to see the whole map regardless of shroud and fog.
-		update_gui_to_player(gui_->viewing_team(), true);
+		update_gui_to_player(static_cast<int>(gui_->viewing_team()), true);
 	}
 
 	bool quit;
