+++ src/statistics.cpp
@@ -224,8 +224,8 @@ static stats::hitrate_map read_by_cth_map_from_battle_result_maps(const statisti
 		for(const auto& j : frequency_map) {
 			const std::string& res = j.first; // see attack_context::~attack_context()
 			const int occurrences = j.second;
-			unsigned int misses = std::count(res.begin(), res.end(), '0');
-			unsigned int hits = std::count(res.begin(), res.end(), '1');
+			long misses = std::count(res.begin(), res.end(), '0');
+			long hits = std::count(res.begin(), res.end(), '1');
 			if(misses + hits == 0) {
 				continue;
 			}
